Also refactor the tests so they are broken out in the same way.
